move FiatIcon & TokenIcon
