+++ src/components/display/FiatIcon.js
