+++ src/components/display/TokenIcon.js
